refactor: Refactor icon component imports to use index file
The icon files have been updated to use an index file for component imports. The import paths for the Props type in icon components have been changed to "./index" from "../icons". Also, the icon export statements in the icon index file have been updated to reflect this change.
