--- src/components/interface/icons.tsx
@@ -1,16 +1,0 @@
-import type { ClassList } from "@builder.io/qwik";
-
-export { FacebookIcon } from "./icons/facebook-icon";
-export { FolderIcon } from "./icons/folder-icon";
-export { GamepadIcon } from "./icons/gamepad-icon";
-export { GitHubIcon } from "./icons/github-icon";
-export { MailIcon } from "./icons/mail-icon";
-export { MarkdownIcon } from "./icons/markdown-icon";
-export { PhoneIcon } from "./icons/phone-icon";
-export { TerminalIcon } from "./icons/terminal-icon";
-export { UserIcon } from "./icons/user-icon";
-export { XIcon } from "./icons/x-icon";
-
-export type Props = {
-	class?: ClassList;
-};
