+++ src/components/interface/icons/markdown-icon.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
 
-import type { Props } from "../icons";
+import type { Props } from "./index";
 
 /** @package */
 export const MarkdownIcon = component$<Props>(({ class: c }) => {
